requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
aiohttp==3.9.5